requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-dotenv>=0.19.0
python-multipart>=0.0.5
psycopg2-binary>=2.9.1
# MQTT dependencies
gmqtt>=0.6.11
asyncio-mqtt>=0.12.1
